backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
cryptography==44.0.3
python-dotenv==1.1.0
Pillow==10.2.0 